+++ src/sim/event_set.cc
@@ -32,7 +32,7 @@ Property* EventSet::newProperty(const std::string& cat, const std::string& nom, 
 }
 
 
-Object * EventSet::newObject(const ObjectTag tag, unsigned num)
+Object * EventSet::newObject(const ObjectTag tag, size_t num)
 {
     Event * e = nullptr;
     if ( tag == Event::TAG )
